tests/submission/DavidCarlisle/math-render01.html

<!DOCTYPE html>
<html>
<head>
<title>math in html: rendered size</title>
<script src="/resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<h1>math in html: rendered size</h1>



<div id="log" style="display:block"></div>

<div style="visibility:hidden">
<span id="s1"><span><span>xxxx</span><span>yyyyyyyyyy</span></span></span>
<span id="s2"><math><mfrac><mi>xxxx</mi><mi>yyyyyyyyyy</mi></mfrac></math></span>
<span id="s3"><math><munder><mi>xxxx</mi><mi>yyyyyyyyyy</mi></munder></math></span>
</div>

<script>


test(function() {
assert_true(document.getElementById("s2").offsetWidth < document.getElementById("s1").offsetWidth)
},"mfrac stacks vertically");

test(function() {
assert_true(document.getElementById("s3").offsetWidth < document.getElementById("s1").offsetWidth)
},"munder stacks vertically");

</script>

           

